+++ src/js/AccessibilityUtils.js
@@ -577,6 +577,43 @@ axs.utils.hasLabel = function(element) {
 };
 
 /**
+ * Determine if this element natively supports being disabled (i.e. via the `disabled` attribute.
+ * Disabled here means that the element should be considered disabled according to specification.
+ * This element may or may not be effectively disabled in practice as this is dependent on implementation.
+ *
+ * @param {Element} element An element to check.
+ * @return {boolean} true If the element supports being natively disabled.
+ */
+axs.utils.isNativelyDisableable = function(element) {
+    var tagName = element.tagName.toUpperCase();
+    return (tagName in axs.constants.NATIVELY_DISABLEABLE);
+};
+
+/**
+ * Determine if this element is disabled directly or indirectly by a disabled ancestor.
+ * Disabled here means that the element should be considered disabled according to specification.
+ * This element may or may not be effectively disabled in practice as this is dependent on implementation.
+ *
+ * @param {Element} element An element to check.
+ * @return {boolean} true if the element or one of its ancestors is disabled.
+ */
+axs.utils.isElementDisabled = function(element) {
+    if (axs.browserUtils.matchSelector(element, '[aria-disabled=true], [aria-disabled=true] *')) {
+        return true;
+    }
+    if(!axs.utils.isNativelyDisableable(element) ||
+            axs.browserUtils.matchSelector(element, 'fieldset>legend:first-of-type *')) {
+        return false;
+    }
+    for (var next = element; next !== null; next = axs.utils.parentElement(next)) {
+        if(axs.utils.isNativelyDisableable(next) && next.hasAttribute('disabled')) {
+            return true;
+        }
+    }
+    return false;
+};
+
+/**
  * @param {Element} element An element to check.
  * @return {boolean} True if the element is hidden from accessibility.
  */
@@ -1119,4 +1156,3 @@ axs.utils.findDescendantsWithRole = function(element, role) {
     }
     return result;
 };
-