+++ Changelog.md
@@ -18,6 +18,7 @@
 * Fix link with clear purpose with text alternative (#156);
 * Handle edge cases in number parser, e.g. "+1", ".1", "01"
 * HTML button containing img with alt attribute now passes controlsWithoutLabel (#202)
+* Disabled elements should be ignored by low contrast audit (#205)
 
 ## 2.8.0 - 2015-07-24